patched kaitai generated files to resolve undefined behavior.
If they are ever re-generated they will likely need to be patched again.

+++ include/kaitai/structs/vlq.h
@@ -113,7 +113,8 @@ public:
     /**
      * Resulting unsigned value as normal integer
      */
-    int32_t value();
+    // patched manually to resolve undefined behavior
+    int64_t value();
 
 private:
     bool f_sign_bit;
